ENGR00310794 iMX6 Fix potential plugin bug which wrecks bootROM register
The R5 register used plugin code is not saved on stack. If this register
is also used in boot ROM, this may cause to destroy the R5 value.

Currently, it is a potential problem because the ROM does not use R5 before
its calling plug-in. But this is the compiler behavior and not be guaranteed.

Signed-off-by: Ye.Li <B37916@freescale.com>