edm-imx8mq: revert the workaround 'force to detect 4GB LPDDR4 as 3GB LPDDR4'
The GPU crash issue is fixed for i.mx8m in linux imx_4.14.98_2.0.0_ga kernel.
So, we revert the workaround below:
-----------------------------------------------------------------------------------------------------------------
It's a temporary workaround to avoid kernel panic in GPU driver, and it's especially for the SOM with 4GB LPDDR4.
SPL still detects DDR size as 4GB, but we force to detect the size as 3GB in u-boot and kernel.
This issue will be fixed in the next version of official NXP BSP release.
-----------------------------------------------------------------------------------------------------------------