axon-imx6: Fix USB host
Fix USB type checking routine to report that the host port 1
is always to be initialized as a host. The OTG port 0 will
check the OTG_ID pin and report that it should be initialized
as either a device or a host.

Previously, these ports were always initialized as USB devices
so using port 1 as a USB host port (via the usb start command)
was not possible.